Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 int main()
 {
-    float i,j,k,l;
-    scanf("%f",&i);
-    scanf("%f",&j);
-    scanf("%f",&k);
-    
-    l= (i+j+k)/ 3.0;
+    int i,j,k,l;
+    scanf("%i",&i);
+    scanf("%i",&j);
+    scanf("%i",&k);
+  
+    l= (i+j+k)/3.0;
     printf("Average: %.2f",&l);
 }
